ai-service/requirements.txt
# FastAPI and server
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart>=0.0.20
pydantic==2.9.2
pydantic-settings==2.6.0

# Vector Database
qdrant-client==1.12.0

# Embeddings and AI
langchain==0.3.7
langchain-community==0.3.5
langchain-nomic==0.1.3
langchain-qdrant==0.2.0
nomic>=3.1.3
openai==1.54.4

# Document Processing
pypdf==5.1.0
pdfplumber==0.11.4
python-docx==1.1.2
pillow>=10.4.0
unstructured==0.16.9
python-magic==0.4.27
# mistralai==1.2.4

# Text Processing and NLP
spacy==3.8.2

# HTTP and async
httpx==0.27.2
aiofiles==24.1.0
requests>=2.32.4

# Utilities
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.3

# Security
python-jose[cryptography]>=3.4.0
passlib[bcrypt]==1.7.4

# Logging and monitoring
loguru==0.7.2
